# btf: use append-style encoding when marshaling