… internal/buffer.go