@lmb 2025-04-22
Overall I'm a bit confused by the outcome of this PR. I'd have thought that dropping allocations this much would give a good boost to runtime, but it almost has a negative effect. Can't figure it out. Maybe one of you has an idea?